requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pandas
openpyxl
python-dotenv
